+++ src/core/actions/sendDIMOTokens.ts
@@ -1,8 +1,10 @@
 import { Chain, Transport, encodeFunctionData } from "viem";
-import { ContractType, ENVIRONMENT, SendDIMOTokens } from "../types/interface.js";
-import { CHAIN_ABI_MAPPING, ENV_MAPPING, SEND_DIMO_TOKENS } from ":core/constants.js";
+import { ContractType, ENVIRONMENT } from "../types/dimoTypes.js";
+import { SEND_DIMO_TOKENS } from ":core/constants/methods.js";
 import { KernelAccountClient, KernelSmartAccount } from "@zerodev/sdk";
 import { EntryPoint } from "permissionless/types";
+import { CHAIN_ABI_MAPPING, ENV_MAPPING } from ":core/constants/mappings.js";
+import { SendDIMOTokens } from ":core/types/args.js";
 
 export async function sendDIMOTokensCallData(
   args: SendDIMOTokens,
