+++ src/core/constants/methods.ts
@@ -1,0 +1,4 @@
+// Contract Methods
+export const MINT_VEHICLE_WITH_DEVICE_DEFINITION = "mintVehicleWithDeviceDefinition";
+export const SET_PERMISSIONS_SACD = "setPermissions";
+export const SEND_DIMO_TOKENS = "transfer";
