+++ src/core/constants/sacdPerms.ts
@@ -1,0 +1,3 @@
+// SACD Permissions
+// Binary: 111111111100 --> Decimal: 4091
+export const SACD_DEFAULT_PERMISSIONS = BigInt(4091);
