+++ src/index.ts
@@ -1,9 +1,7 @@
-export type {
-  MintVehicleWithDeviceDefinition,
-  SendDIMOTokens,
-  SetVehiclePermissions,
-  ENVIRONMENT,
-} from "./core/types/interface.js";
+export type { ENVIRONMENT } from "./core/types/dimoTypes.js";
+export type { MintVehicleWithDeviceDefinition, SendDIMOTokens, SetVehiclePermissions } from "./core/types/args.js";
+export type { VehicleNodeMintedWithDeviceDefinition } from "./core/types/eventLogs.js";
+export { ContractType } from "./core/types/dimoTypes.js";
 export {
   mintVehicleWithDeviceDefinition,
   mintVehicleCallData,
