+++ src/version.ts
@@ -1,1 +1,1 @@
-export const LIB_VERSION = '1.2.0';
+export const LIB_VERSION = '1.2.02';
